+++ .idea/.name
@@ -1,1 +1,1 @@
-aiq.android
+aiq.components
