+++ aiqcomponents/build.gradle
