+++ aiqcomponents/src/main/res/drawable/ic_default_icon.xml
