+++ aiqcomponents/src/main/res/drawable/ic_search_grey.xml
