+++ aiqcomponents/src/main/res/drawable/rounded_search_shape.xml
