+++ aiqcomponents/src/main/res/layout/selector.xml
