+++ aiqcomponents/src/main/res/layout/text_input.xml
