+++ aiqcomponents/src/main/res/values/attrs.xml
