+++ aiqcomponents/src/main/res/values/colors.xml
