+++ aiqcomponents/src/main/res/values/dimens.xml
