+++ aiqcomponents/src/main/res/values/strings.xml
