+++ aiqcomponents/src/main/res/values/styles.xml
