--- aiqinput/src/main/res/layout/itemlist_icon_sub_title.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <data>
-        <variable
-            name="viewObject"
-            type="com.aiqfome.aiqinput.adapters.IconItem" />
-    </data>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <View
-            android:id="@+id/top_divider"
-            android:background="@color/colorDivider"
-            android:layout_width="match_parent"
-            android:layout_height="1dp" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="?android:attr/selectableItemBackground"
-            android:gravity="center_vertical"
-            android:orientation="horizontal"
-            android:paddingTop="16dp"
-            android:paddingBottom="16dp">
-
-            <ImageView
-                android:id="@+id/iv_icon"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="25dp"
-                android:layout_marginEnd="20dp"
-                app:srcCompat="@drawable/ic_default_icon" />
-
-            <TextView
-                android:id="@+id/tv_subtitle"
-                style="@style/TextAppearance.aiq.BoldBody"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginEnd="10dp"
-                android:text="@{viewObject.subTitle}" />
-
-            <TextView
-                android:id="@+id/tv_title"
-                style="@style/TextAppearance.aiq.BoldBody"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@{viewObject.title}" />
-
-        </LinearLayout>
-
-    </LinearLayout>
-
-</layout>
