--- aiqinput/src/main/res/layout/itemlist_sub_title.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <data>
-        <variable
-            name="viewObject"
-            type="com.aiqfome.aiqinput.adapters.BaseItem" />
-    </data>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <View
-            android:id="@+id/top_divider"
-            android:background="@color/colorDivider"
-            android:layout_width="match_parent"
-            android:layout_height="1dp" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="?android:attr/selectableItemBackground"
-            android:gravity="center_vertical"
-            android:orientation="horizontal"
-            android:paddingStart="20dp"
-            android:paddingTop="16dp"
-            android:paddingBottom="16dp">
-
-            <TextView
-                android:id="@+id/tv_subtitle"
-                style="@style/TextAppearance.aiq.BoldBody"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginEnd="10dp"
-                android:text="@{viewObject.subTitle}" />
-
-            <TextView
-                android:id="@+id/tv_title"
-                style="@style/TextAppearance.aiq.BoldBody"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@{viewObject.title}" />
-
-        </LinearLayout>
-
-    </LinearLayout>
-
-</layout>
