+++ settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name='aiq.android'
+rootProject.name='aiq.components'
 include ':app'
-include ':aiqinput'
+include ':aiqcomponents'
